refactor(macros): extract namespace parts to a dedicated file
